Finish functional hoc
