+++ src/components/Select/MultiSelect/hooks.js
@@ -1,27 +1,31 @@
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import { usePrevious } from "hooks";
 import {
   useHandleKeyboardHook,
-  useHandleKeyboardHookBasicParameters,
-  useSelect
+  useHandleKeyboardHookBasicParameters
 } from "../hooks";
-import { matchOption, findExact } from "../helpers";
+import { matchOption } from "../helpers";
 
 export function useMultiSelect(
   disabled,
-  autoFocus,
   onChange,
   options,
   getOptionLabel,
-  getOptionValue,
   optionsFilter,
   searchable,
   value,
   inputValue,
-  onInputChange
-) {
-  const [_options, setOptions] = useState([]);
+  onInputChange,
+  _options,
+  setOptions,
 
+  focusedOptionIndex,
+  setFocusedOptionIndex,
+  menuOpened,
+  selectOption,
+  resetMenu,
+  removeSelectedOptionFilter
+) {
   useEffect(() => {
     let filteredOptions = optionsFilter
       ? options.filter(optionsFilter)
@@ -33,7 +37,14 @@ export function useMultiSelect(
         inputValue
       );
     setOptions(filteredOptions);
-  }, [searchable, getOptionLabel, inputValue, optionsFilter, options]);
+  }, [
+    searchable,
+    getOptionLabel,
+    inputValue,
+    optionsFilter,
+    options,
+    setOptions
+  ]);
 
   const previousSelectedOptions = usePrevious(value);
 
@@ -43,59 +54,10 @@ export function useMultiSelect(
       onChange(value.filter(optionsFilter));
   }, [disabled, value, optionsFilter, previousSelectedOptions, onChange]);
 
-  const removeSelectedOptionFilter = (option) =>
-    value.filter(
-      (selectedOption) =>
-        getOptionLabel(selectedOption) !== getOptionLabel(option)
-    );
-
   const removeSelectedOption = (e, option) => {
     onChange(removeSelectedOptionFilter(option));
     e.stopPropagation();
   };
-
-  const setOption = (option, knownIndex) => {
-    if (disabled) return;
-    const index =
-      knownIndex || findExact(_options, getOptionLabel, getOptionValue, option);
-    let newSelectedOptions = [];
-    if (
-      value.find(
-        (selectedOption) =>
-          getOptionLabel(selectedOption) === getOptionLabel(option)
-      )
-    )
-      newSelectedOptions = removeSelectedOptionFilter(option);
-    else if (!value.length) {
-      newSelectedOptions = [option];
-    } else {
-      newSelectedOptions = [...value, option];
-    }
-    setFocusedOptionIndex(index);
-    if (searchable && onInputChange && inputValue) onInputChange("");
-    onChange(newSelectedOptions);
-  };
-
-  const {
-    focusedOptionIndex,
-    setFocusedOptionIndex,
-    menuOpened,
-    selectOption,
-    openMenu,
-    containerRef,
-    resetMenu,
-    onSearch,
-    transitions
-  } = useSelect(
-    _options,
-    setOption,
-    disabled,
-    onInputChange,
-    inputValue,
-    autoFocus,
-    onChange,
-    searchable
-  );
 
   const {
     onTypeArrowDownHandler,
@@ -126,16 +88,7 @@ export function useMultiSelect(
   };
 
   return {
-    _options,
-    containerRef,
-    menuOpened,
-    focusedOptionIndex,
-    openMenu,
-    setFocusedOptionIndex,
-    selectOption,
     cancelSelection,
-    removeSelectedOption,
-    onSearch,
-    transitions
+    removeSelectedOption
   };
 }
