+++ src/index.js
@@ -40,6 +40,7 @@ export { default as P } from "./components/Typography/P.jsx";
 export { default as Text } from "./components/Typography/Text/Text.jsx";
 export { default as Badge } from "./components/Badge/Badge.jsx";
 export { default as Toggle } from "./components/Toggle/Toggle.jsx";
+export { default as SelectWrapper } from "./components/Select/SelectWrapper.jsx";
 export { default as Select } from "./components/Select/BasicSelect/BasicSelect.jsx";
 export { default as Creatable } from "./components/Select/CreatableSelect/CreatableSelect.jsx";
 export { default as MultiSelect } from "./components/Select/MultiSelect/MultiSelect.jsx";
